Apply Codex security review fixes
- Enforce mandatory signer authorization (fail-closed without auth table)
- Treat X-Skip-Signing/X-Signer as untrusted and do not honor skip requests
- Restrict milter network sockets to loopback; harden Unix socket creation
- Redact Redis URI from process arguments after parsing
- Harden Redis: size limits for PEM/chain, HSTRLEN-before-fetch, NUL handling,
  explicit TLS hostname verification, and reject verify=none/plaintext
- Fix multipart/signed parsing and semicolon header handling
- Limit Content-* header chain size and append cost
- Add recipient counting for skip-mode modetable
- Update unit and integration tests; install python3-miltertest via uv

+++ src/utils/get_num_semicolons.c
@@ -5,18 +5,16 @@
  *
  * argument: pointer to a null-terminated string
  * return:   number of semicolons
- *           -1 if the argument is a NULL pointer
- * bugs    : integer overflow if more than sizeof(int)/2 - 1
- *           semicolons occur in the string
+ *           (size_t)-1 if the argument is a NULL pointer
  */
-int get_num_semicolons(const char* string) {
+size_t get_num_semicolons(const char* string) {
 
-    int num_semicolon = 0;
+    size_t      num_semicolon = 0;
     const char* p;
 
     if (string == NULL) {
         logmsg(LOG_ERR, "FATAL: get_num_semikolons failed: got empty string");
-        return(-1);
+        return (size_t)-1;
     }
 
     p = string;
@@ -25,6 +23,6 @@ int get_num_semicolons(const char* string) {
             num_semicolon++;
         p++;
     }
-    logmsg(LOG_DEBUG, "get_num_semicolons: %i Semicolon in ->%s<-", num_semicolon, string);
-    return(num_semicolon);
+    logmsg(LOG_DEBUG, "get_num_semicolons: %zu Semicolon in ->%s<-", num_semicolon, string);
+    return num_semicolon;
 }
